project-2/requirements.txt
pandas>=1.5.0
matplotlib>=3.5.0
statsmodels>=0.13.0
streamlit>=1.29.0
plotly>=5.0.0

# Additional dependencies for the Airflow/Kafka pipeline
kafka-python>=2.0.0
psycopg2-binary>=2.9.0
requests>=2.28.0